Try to write the TLS top-level interface (Adding TLS.fs file).
In the meanwhile, try to better handle re-negotiation corner cases.
This leads to the rewriting of part of the dispatcher code and handshake and other protocols.

Add the moment, the top level TLS interface is only partly implemented, while dispatch
and protocols should be reasonably working (untested). Handshake is in a very preliminary state,
but the interface should be reasonable.

All fs7 interfaces are currently misaligned, and need to be ported to this later version of the implementation.