- Try to improve session resumption.
At the moment we get "inconsisten" behavior that must be further investigated. Sometimes decryption at record layer gives back "strange" data (wrong padding at least) which makes everything fail -- maybe ms is badly generated on resumption?. Also, when we ask for resumption, but we get a full handshake, it seems that the user is not properly notified.

- Document a bug in dispatcher (by a FIXME comment): Sometimes dispatcher tells the user some app_data have been sent, while they have not been sent in fact.

- Add some user-interaction function for sessionDB.